nastcap.core/pyproject.toml
[tool.poetry]
name = "nation_gui"
version = "1.0"
description = ""
authors = ["Taylor B. <43686206+tayjaybabee@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.8, <=3.10"
nationstates = "^3.1.4"
inspy-logger = "2.1a14"
PySimpleGUI = "^4.55.1"
python-box = "^5.4.1"
wget = "^3.2"
clipboard = "^0.0.4"
inspyre-toolbox = "^1.2.1"
pynput = "^1.7.6"
Pillow = "^9.0.0"
matplotlib = "^3.5.1"
pandas = "^1.4.1"
seaborn = "^0.11.2"
asciimatics = "^1.13.0"
PySimpleGUIQt = "^0.35.0"

[tool.poetry.dev-dependencies]
ptipython = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
